+++ example/lib/examples/other_buttons_example.dart
@@ -10,7 +10,6 @@ class _OtherButtonExampleState extends State<OtherButtonExample> {
   final isLoading = ValueNotifier<bool>(false);
 
   SuraLoadingDialog? loadingDialog;
-  FutureManager? futureManager;
 
   Future onButtonClick() async {
     await Future.delayed(Duration(seconds: 2));
@@ -131,8 +130,7 @@ class _OtherButtonExampleState extends State<OtherButtonExample> {
     );
   }
 
-  Widget buildButtonCategory(
-      {required String name, required List<Widget> buttons}) {
+  Widget buildButtonCategory({required String name, required List<Widget> buttons}) {
     return Column(
       children: [
         Text(name, style: TextStyle(fontSize: 16, fontWeight: FontWeight.w500)),