+++ lib/src/style/shadow_input_border.dart
@@ -11,7 +11,7 @@ class ShadowInputBorder extends InputBorder {
   ///Color of the shadow
   final Color shadowColor;
 
-  ///COlor that fill the entired TextField
+  ///Color that fill the entire TextField
   final Color fillColor;
 
   /// Creates an shadow border for an [InputDecorator]
@@ -34,11 +34,7 @@ class ShadowInputBorder extends InputBorder {
 
   @override
   Path getInnerPath(Rect rect, {TextDirection? textDirection}) {
-    return Path()
-      ..addRRect(borderRadius
-          .resolve(textDirection)
-          .toRRect(rect)
-          .deflate(borderSide.width));
+    return Path()..addRRect(borderRadius.resolve(textDirection).toRRect(rect).deflate(borderSide.width));
   }
 
   @override