+++ lib/src/style/small_underline_tab_indicator.dart
@@ -19,7 +19,7 @@ class SmallUnderLineTabIndicator extends Decoration {
   ///indicator color
   final Color color;
 
-  ///[DotAlignment] wether it's bottom or top of tab's label
+  ///[DotAlignment] whether it's bottom or top of tab's label
   final TabAlignment tabAlignment;
 
   SmallUnderLineTabIndicator({
@@ -63,9 +63,7 @@ class _SmallUnderLineTabIndicatorPainter extends BoxPainter {
   ) : super(onChanged);
   @override
   void paint(Canvas canvas, Offset offset, ImageConfiguration configuration) {
-    final double yOffset = tabAlignment == TabAlignment.bottom
-        ? configuration.size!.height - 8
-        : 0;
+    final double yOffset = tabAlignment == TabAlignment.bottom ? configuration.size!.height - 8 : 0;
     final Offset indicatorOffset = offset + Offset(paddingLeft, yOffset);
 
     final paint = Paint()