+++ lib/src/widgets/sura_async_button.dart
@@ -47,7 +47,7 @@ class SuraAsyncButton extends StatefulWidget {
   ///Button's shape
   final OutlinedBorder? shape;
 
-  ///button's borderside
+  ///button's border side
   final BorderSide? borderSide;
 
   ///Button's elevation
@@ -129,9 +129,8 @@ class _SuraAsyncButtonState extends State<SuraAsyncButton> {
       ],
     );
 
-    final Widget loadingWidget = widget.loadingWidget ??
-        SuraProvider.of(context)?.buttonLoadingWidget ??
-        _buildLoadingWidget();
+    final Widget loadingWidget =
+        widget.loadingWidget ?? SuraProvider.of(context)?.buttonLoadingWidget ?? _buildLoadingWidget();
 
     return Container(
       height: widget.height,
@@ -147,9 +146,7 @@ class _SuraAsyncButtonState extends State<SuraAsyncButton> {
                 : null,
         child: ConditionalWidget(
           condition: isLoading,
-          onTrue: () => widget.loadingType == LoadingType.Disable
-              ? buttonContent
-              : loadingWidget,
+          onTrue: () => widget.loadingType == LoadingType.Disable ? buttonContent : loadingWidget,
           onFalse: () => buttonContent,
         ),
         style: ElevatedButton.styleFrom(