+++ lib/src/widgets/sura_paginated_list.dart
@@ -74,8 +74,8 @@ class _SuraPaginatedListState extends State<SuraPaginatedList> {
     if (loadingState > 1) return;
     if (widget.hasMoreData) {
       await widget.dataLoader.call();
-      if(mounted){
-        loadingState = 0;      
+      if (mounted) {
+        loadingState = 0;
       }
     }
   }
@@ -83,10 +83,9 @@ class _SuraPaginatedListState extends State<SuraPaginatedList> {
   void initController() {
     if (_isPrimaryScrollView) {
       scrollController = ScrollController();
-      scrollController.addListener(() => scrollListener(scrollController));
+      scrollController!.addListener(() => scrollListener(scrollController!));
     } else {
-      widget.scrollController
-          ?.addListener(() => scrollListener(widget.scrollController!));
+      widget.scrollController?.addListener(() => scrollListener(widget.scrollController!));
     }
   }